all codes arranged
