+++ api/models.py
@@ -4,20 +4,6 @@ from django.contrib.auth.models import User
 
 # Create your models here.
 
-# class User(models.Model):
-#     # firstName = models.CharField(max_length=50,null=False,blank=False)
-#     # lastName = models.CharField(max_length=50,null=False,blank=False)
-#     username = models.CharField(max_length=50,null=False,blank=False)
-#     email = models.EmailField(null=False,blank=False)
-#     password = models.CharField(max_length=20,null=False,blank=False)
-#     usertype = models.CharField(max_length=50,null=False,blank=False)
-#     phone = models.CharField(max_length=50,null=True,blank=True)
-#     image = models.ImageField(upload_to = 'uploads/images/userImages',null=False,blank=False,default='')
-#     def __str__(self):
-#         return self.username
-
-
-
 class Address(models.Model):
     province = models.CharField(max_length=20)
     def __str__(self):
