+++ frontend/src/Components/user/Logout.js
@@ -1,7 +1,9 @@
-const Logout = () => {
-  localStorage.removeItem("token");
-  window.location = "/signin/:id/";
+import { useNavigate } from "react-router-dom";
 
+const Logout = () => {
+  const navigat = useNavigate()
+  localStorage.removeItem("token");
+  navigat('/signin/')
   return null;
 };
